Writing final currentTime, sunrise, & sunset with .toInt()...w

+++ lib/models/current_weather.dart
@@ -114,9 +114,9 @@ class CurrentWeather {
     final city = data['name'];
     final description = data['weather'][0]['description'];
     final currentTemp = data['main']['temp'].toDouble();
-    final currentTime = DateTime.fromMillisecondsSinceEpoch(data['dt'] * 1000, isUtc: true).toLocal();
-    final sunrise = DateTime.fromMillisecondsSinceEpoch(data['sys']['sunrise'] * 1000, isUtc: true).toLocal();
-    final sunset = DateTime.fromMillisecondsSinceEpoch(data['sys']['sunset'] * 1000, isUtc: true).toLocal();
+    final currentTime = DateTime.fromMillisecondsSinceEpoch((data['dt'] * 1000).toInt(), isUtc: true).toLocal();
+    final sunrise = DateTime.fromMillisecondsSinceEpoch((data['sys']['sunrise'] * 1000).toInt(), isUtc: true).toLocal();
+    final sunset = DateTime.fromMillisecondsSinceEpoch((data['sys']['sunset'] * 1000).toInt(), isUtc: true).toLocal();
 
     return CurrentWeather(
       city: city,
